SNC-5478. Low disk space displaying wrong size.

Closes SNC-5478

See merge request desktop/desktop!3570

+++ src/MEGASync/gui/LowDiskSpaceDialog.cpp
@@ -6,7 +6,7 @@
 
 #include <QFileIconProvider>
 
-LowDiskSpaceDialog::LowDiskSpaceDialog(long neededSize,
+LowDiskSpaceDialog::LowDiskSpaceDialog(long long neededSize,
                                        long long freeSize,
                                        long long driveSize,
                                        const QString& driveName,
